requirements.txt
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.1
executing==1.1.1
idna==3.4
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
lxml==4.9.1
matplotlib-inline==0.1.6
numpy==1.23.4
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.1
toml==0.10.2
traitlets==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
